backend/requirements.txt
Flask
flask-cors
flask-migrate
flask-sqlalchemy
flask-login
flask-jwt-extended
celery
redis
Pillow
nltk
matplotlib
numpy
textblob
wordcloud
requests
PyPDF2
python-docx
pandas
beautifulsoup4
opencv-python
textstat
langdetect
scikit-learn
gensim
waitress